Moved 'counter' related components into components/counts/ for consistency sake. Resolves #38
