+++ src/components/counts/Counter.svelte
