+++ src/components/counts/Protein.svelte
@@ -1,6 +1,6 @@
 <script>
-  import Counter from '../components/Counter.svelte';
-  import { todayStore, settingStore } from '../stores/local';
+  import Counter from '$components/counts/Counter.svelte';
+  import { todayStore, settingStore } from '$stores/local';
   let title = '🍗 protein (g)';
 </script>
