+++ src/components/counts/Salt.svelte
@@ -1,6 +1,6 @@
 <script>
-    import Counter from '../components/Counter.svelte';
-  import { todayStore, settingStore } from '../stores/local';
+    import Counter from '$components/counts/Counter.svelte';
+  import { todayStore, settingStore } from '$stores/local';
     let title = '🧂 sodium (mg)';
 </script>
