+++ src/components/goals/Calories.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Counter from '$components/Counter.svelte';
+    import Counter from '$components/counts/Counter.svelte';
     import { limitStore } from '$stores/local';
     let title = '⚡ calories';
 </script>
